feat: changed layout to add sidebar
