+++ apps/www/components/ui/resizable.tsx
@@ -1,3 +1,4 @@
+"use client";
 import { DragHandleDots2Icon } from "@radix-ui/react-icons";
 import * as ResizablePrimitive from "react-resizable-panels";
